work on rate limit
Included reporting on the Strava rate limit. Added a function to count
the available pages, but it actually reports one page to many. Probably
the offset of the page counting should be something intelligent (e.g.
in the properties), as there is no way to get the number of records
from Strava